fix: rush cov solid
